+++ ChangeLog.md
@@ -1,3 +1,15 @@
+# 0.30
+- add TargetFrameworks: net6.0-windows, net452
+- remove TargetFrameworks: net45
+
+# 0.21
+- EventManager.Raise add overload with lazy event args
+- FIX ActionVM.RegisterActionMethod
+- FIX do not RegisterActionMethod if Metadata.ActionProvider.ExecutedCallback is already set
+- add ButtonActionVM. derived from ActionVM supporting common button properties
+- add some UIProperties (ObjectVM.UI)
+- FIX ListVM clear item.Parent when removed
+
 # 0.20
 - BREAKING CHANGE ValueChangedEventArgs swap arguments and rename PreviousValue to OldValue, so that you get a compiler error on event handlers. If you create ValueChangedEventArgs you have to swap the arguments manually. No Warning in this case!
 - add SimpleCommand.RaiseCanExecuteChanged